#3 refactor: 유효성 검사 로직을 커스텀 훅으로 전환
