+++ pages/signup/index.tsx
@@ -1,8 +1,9 @@
 import classNames from 'classnames';
 import styles from './signup.module.scss';
 import Link from 'next/link';
-import { SubmitHandler, useForm } from 'react-hook-form';
+import { SubmitHandler } from 'react-hook-form';
 import React from 'react';
+import { useValidForm } from '@/hooks/useValidForm';
 
 interface FormInputValues {
   name: string;
@@ -12,51 +13,12 @@ interface FormInputValues {
 }
 
 export default function SignupPage() {
-  const {
-    register,
-    trigger,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<FormInputValues>({ mode: 'onChange' });
+  const { register, errors, handleSubmit } = useValidForm();
 
   const handleFormSubmit: SubmitHandler<FormInputValues> = data => {
     // await fetch 회원가입
     console.log(data);
   };
-
-  const nameRegister = register('name', {
-    required: true,
-    minLength: { value: 2, message: '2자 이상으로 작성해주세요.' },
-    maxLength: { value: 10, message: '10자 이내로 작성해주세요.' },
-  });
-
-  const emailRegister = register('email', {
-    required: true,
-    pattern: {
-      value: /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/,
-      message: '이메일 형식으로 작성해주세요',
-    },
-  });
-
-  const passwordRegister = register('password', {
-    required: true,
-    pattern: {
-      value: /^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$/,
-      message: '영문, 숫자를 포함하여 8자 이상으로 작성해주세요',
-    },
-    onChange: () => trigger('verifyPassword'),
-  });
-
-  const verifyPasswordRegister = register('verifyPassword', {
-    required: true,
-    pattern: {
-      value: /^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$/,
-      message: '영문, 숫자를 포함하여 8자 이상으로 작성해주세요',
-    },
-    validate: {
-      matched: (value, formValues) => value === formValues.password || '비밀번호가 일치하지 않습니다.',
-    },
-  });
 
   return (
     <main className={styles.container}>
@@ -71,7 +33,7 @@ export default function SignupPage() {
             id={'name'}
             className={classNames('input', errors.name && styles.error)}
             placeholder={'이름을 입력해 주세요'}
-            {...nameRegister}
+            {...register.name}
           />
           {errors.name && <p className={classNames('text-xs', styles.message)}>{errors.name.message}</p>}
         </div>
@@ -85,7 +47,7 @@ export default function SignupPage() {
             type={'email'}
             className={classNames('input', errors.email && styles.error)}
             placeholder={'이메일을 입력해 주세요'}
-            {...emailRegister}
+            {...register.email}
           />
           {errors.email && <p className={classNames('text-xs', styles.message)}>{errors.email.message}</p>}
         </div>
@@ -98,7 +60,7 @@ export default function SignupPage() {
             id={'password'}
             className={classNames('input', errors.password && styles.error)}
             placeholder={'비밀번호를 입력해 주세요'}
-            {...passwordRegister}
+            {...register.password}
           />
           {errors.password && <p className={classNames('text-xs', styles.message)}>{errors.password.message}</p>}
         </div>
@@ -111,7 +73,7 @@ export default function SignupPage() {
             id={'verifyPassword'}
             className={classNames('input', errors.verifyPassword && styles.error)}
             placeholder={'비밀번호를 입력해 주세요'}
-            {...verifyPasswordRegister}
+            {...register.verifyPassword}
           />
           {errors.verifyPassword && (
             <p className={classNames('text-xs', styles.message)}>{errors.verifyPassword.message}</p>
